+++ demo/libwke/wke.h
@@ -14,13 +14,6 @@
 //////////////////////////////////////////////////////////////////////////
 
 
-#define WIN32_LEAN_AND_MEAN
-#include <windows.h>
-
-
-//////////////////////////////////////////////////////////////////////////
-
-
 #if defined(__cplusplus)
     #define WKE_EXTERN_C extern "C"
 #else
@@ -29,10 +22,16 @@
 
 
 #if defined(BUILDING_wke)
-#   define WKE_API WKE_EXTERN_C __declspec(dllexport)
+#   define WKE_SYMBOL __declspec(dllexport)
 #else
-#   define WKE_API WKE_EXTERN_C __declspec(dllimport)
+#   define WKE_SYMBOL __declspec(dllimport)
 #endif
+
+
+#define WKE_CALL _cdecl
+#define WKE_API  WKE_EXTERN_C WKE_SYMBOL
+
+
 
 
 typedef struct 
@@ -162,149 +161,149 @@ extern "C"
 {
 #endif
 
-WKE_API void wkeInitialize();
-WKE_API void wkeInitializeEx(const wkeSettings* settings);
-WKE_API void wkeConfigure(const wkeSettings* settings);
+WKE_API void WKE_CALL wkeInitialize();
+WKE_API void WKE_CALL wkeInitializeEx(const wkeSettings* settings);
+WKE_API void WKE_CALL wkeConfigure(const wkeSettings* settings);
 
-WKE_API void wkeFinalize();
-WKE_API void wkeUpdate();
-WKE_API unsigned int wkeGetVersion();
-WKE_API const utf8* wkeGetVersionString();
+WKE_API void WKE_CALL wkeFinalize();
+WKE_API void WKE_CALL wkeUpdate();
+WKE_API unsigned int WKE_CALL wkeGetVersion();
+WKE_API const utf8* WKE_CALL wkeGetVersionString();
 
-typedef void* (*FILE_OPEN) (const char* path);
-typedef void (*FILE_CLOSE) (void* handle);
-typedef size_t (*FILE_SIZE) (void* handle);
-typedef int (*FILE_READ) (void* handle, void* buffer, size_t size);
-typedef int (*FILE_SEEK) (void* handle, int offset, int origin);
+typedef void* (WKE_CALL *FILE_OPEN) (const char* path);
+typedef void (WKE_CALL *FILE_CLOSE) (void* handle);
+typedef size_t (WKE_CALL *FILE_SIZE) (void* handle);
+typedef int (WKE_CALL *FILE_READ) (void* handle, void* buffer, size_t size);
+typedef int (WKE_CALL *FILE_SEEK) (void* handle, int offset, int origin);
 
-WKE_API void wkeSetFileSystem(FILE_OPEN pfn_open, FILE_CLOSE pfn_close, FILE_SIZE pfn_size, FILE_READ pfn_read, FILE_SEEK pfn_seek);
-
-
-WKE_API wkeWebView wkeCreateWebView();
-WKE_API wkeWebView wkeGetWebView(const char* name);
-WKE_API void wkeDestroyWebView(wkeWebView webView);
-
-WKE_API const char* wkeGetName(wkeWebView webView);
-WKE_API void wkeSetName(wkeWebView webView, const char* name);
-
-WKE_API bool wkeIsTransparent(wkeWebView webView);
-WKE_API void wkeSetTransparent(wkeWebView webView, bool transparent);
-
-WKE_API void wkeSetUserAgent(wkeWebView webView, const utf8* userAgent);
-WKE_API void wkeSetUserAgentW(wkeWebView webView, const wchar_t* userAgent);
-
-WKE_API void wkeLoadURL(wkeWebView webView, const utf8* url);
-WKE_API void wkeLoadURLW(wkeWebView webView, const wchar_t* url);
-WKE_API void wkePostURL(wkeWebView wkeView,const utf8* url, const char* postData,int  postLen); 
-WKE_API void wkePostURLW(wkeWebView wkeView,const wchar_t* url, const char* postData, int postLen);
-
-WKE_API void wkeLoadHTML(wkeWebView webView, const utf8* html);
-WKE_API void wkeLoadHTMLW(wkeWebView webView, const wchar_t* html);
-
-WKE_API void wkeLoadFile(wkeWebView webView, const utf8* filename);
-WKE_API void wkeLoadFileW(wkeWebView webView, const wchar_t* filename);
-
-WKE_API void wkeLoad(wkeWebView webView, const utf8* str);
-WKE_API void wkeLoadW(wkeWebView webView, const wchar_t* str);
-
-WKE_API bool wkeIsLoading(wkeWebView webView);
-WKE_API bool wkeIsLoadingSucceeded(wkeWebView webView);
-WKE_API bool wkeIsLoadingFailed(wkeWebView webView);
-WKE_API bool wkeIsLoadingCompleted(wkeWebView webView);
-WKE_API bool wkeIsDocumentReady(wkeWebView webView);
-WKE_API void wkeStopLoading(wkeWebView webView);
-WKE_API void wkeReload(wkeWebView webView);
-
-WKE_API const utf8* wkeGetTitle(wkeWebView webView);
-WKE_API const wchar_t* wkeGetTitleW(wkeWebView webView);
-
-WKE_API void wkeResize(wkeWebView webView, int w, int h);
-WKE_API int wkeGetWidth(wkeWebView webView);
-WKE_API int wkeGetHeight(wkeWebView webView);
-WKE_API int wkeGetContentWidth(wkeWebView webView);
-WKE_API int wkeGetContentHeight(wkeWebView webView);
-
-WKE_API void wkeSetDirty(wkeWebView webView, bool dirty);
-WKE_API bool wkeIsDirty(wkeWebView webView);
-WKE_API void wkeAddDirtyArea(wkeWebView webView, int x, int y, int w, int h);
-WKE_API void wkeLayoutIfNeeded(wkeWebView webView);
-WKE_API void wkePaint(wkeWebView webView, void* bits,int bufWid, int bufHei, int xDst, int yDst, int w, int h, int xSrc, int ySrc, bool bCopyAlpha);
-WKE_API void wkePaint2(wkeWebView webView, void* bits,int pitch);
-WKE_API void wkeRepaintIfNeeded(wkeWebView webView);
-WKE_API HDC wkeGetViewDC(wkeWebView webView);
-
-WKE_API bool wkeCanGoBack(wkeWebView webView);
-WKE_API bool wkeGoBack(wkeWebView webView);
-WKE_API bool wkeCanGoForward(wkeWebView webView);
-WKE_API bool wkeGoForward(wkeWebView webView);
-
-WKE_API void wkeEditorSelectAll(wkeWebView webView);
-WKE_API void wkeEditorCopy(wkeWebView webView);
-WKE_API void wkeEditorCut(wkeWebView webView);
-WKE_API void wkeEditorPaste(wkeWebView webView);
-WKE_API void wkeEditorDelete(wkeWebView webView);
-
-WKE_API const wchar_t* wkeGetCookieW(wkeWebView webView);
-WKE_API const utf8* wkeGetCookie(wkeWebView webView);
-WKE_API void wkeSetCookieEnabled(wkeWebView webView, bool enable);
-WKE_API bool wkeIsCookieEnabled(wkeWebView webView);
-
-WKE_API void wkeSetMediaVolume(wkeWebView webView, float volume);
-WKE_API float wkeGetMediaVolume(wkeWebView webView);
-
-WKE_API bool wkeFireMouseEvent(wkeWebView webView, unsigned int message, int x, int y, unsigned int flags);
-WKE_API bool wkeFireContextMenuEvent(wkeWebView webView, int x, int y, unsigned int flags);
-WKE_API bool wkeFireMouseWheelEvent(wkeWebView webView, int x, int y, int delta, unsigned int flags);
-WKE_API bool wkeFireKeyUpEvent(wkeWebView webView, unsigned int virtualKeyCode, unsigned int flags, bool systemKey);
-WKE_API bool wkeFireKeyDownEvent(wkeWebView webView, unsigned int virtualKeyCode, unsigned int flags, bool systemKey);
-WKE_API bool wkeFireKeyPressEvent(wkeWebView webView, unsigned int charCode, unsigned int flags, bool systemKey);
-
-WKE_API void wkeSetFocus(wkeWebView webView);
-WKE_API void wkeKillFocus(wkeWebView webView);
-
-WKE_API wkeRect wkeGetCaretRect(wkeWebView webView);
-
-WKE_API jsValue wkeRunJS(wkeWebView webView, const utf8* script);
-WKE_API jsValue wkeRunJSW(wkeWebView webView, const wchar_t* script);
-
-WKE_API jsExecState wkeGlobalExec(wkeWebView webView);
-
-WKE_API void wkeSleep(wkeWebView webView);
-WKE_API void wkeWake(wkeWebView webView);
-WKE_API bool wkeIsAwake(wkeWebView webView);
-
-WKE_API void wkeSetZoomFactor(wkeWebView webView, float factor);
-WKE_API float wkeGetZoomFactor(wkeWebView webView);
-
-WKE_API void wkeSetEditable(wkeWebView webView, bool editable);
-
-WKE_API void wkeSetHostWindow(wkeWebView webWindow, HWND hostWindow);
-WKE_API HWND wkeGetHostWindow(wkeWebView webWindow);
+WKE_API void WKE_CALL wkeSetFileSystem(FILE_OPEN pfn_open, FILE_CLOSE pfn_close, FILE_SIZE pfn_size, FILE_READ pfn_read, FILE_SEEK pfn_seek);
 
 
-WKE_API const utf8* wkeGetString(const wkeString string);
-WKE_API const wchar_t* wkeGetStringW(const wkeString string);
+WKE_API wkeWebView WKE_CALL wkeCreateWebView();
+WKE_API wkeWebView WKE_CALL wkeGetWebView(const char* name);
+WKE_API void WKE_CALL wkeDestroyWebView(wkeWebView webView);
 
-WKE_API void wkeSetString(wkeString string, const utf8* str, size_t len);
-WKE_API void wkeSetStringW(wkeString string, const wchar_t* str, size_t len);
+WKE_API const char* WKE_CALL wkeGetName(wkeWebView webView);
+WKE_API void WKE_CALL wkeSetName(wkeWebView webView, const char* name);
 
-typedef void (*wkeTitleChangedCallback)(wkeWebView webView, void* param, const wkeString title);
-WKE_API void wkeOnTitleChanged(wkeWebView webView, wkeTitleChangedCallback callback, void* callbackParam);
+WKE_API bool WKE_CALL wkeIsTransparent(wkeWebView webView);
+WKE_API void WKE_CALL wkeSetTransparent(wkeWebView webView, bool transparent);
 
-typedef void (*wkeURLChangedCallback)(wkeWebView webView, void* param, const wkeString url);
-WKE_API void wkeOnURLChanged(wkeWebView webView, wkeURLChangedCallback callback, void* callbackParam);
+WKE_API void WKE_CALL wkeSetUserAgent(wkeWebView webView, const utf8* userAgent);
+WKE_API void WKE_CALL wkeSetUserAgentW(wkeWebView webView, const wchar_t* userAgent);
 
-typedef void (*wkePaintUpdatedCallback)(wkeWebView webView, void* param, const HDC hdc, int x, int y, int cx, int cy);
-WKE_API void wkeOnPaintUpdated(wkeWebView webView, wkePaintUpdatedCallback callback, void* callbackParam);
+WKE_API void WKE_CALL wkeLoadURL(wkeWebView webView, const utf8* url);
+WKE_API void WKE_CALL wkeLoadURLW(wkeWebView webView, const wchar_t* url);
+WKE_API void WKE_CALL wkePostURL(wkeWebView wkeView,const utf8* url, const char* postData,int  postLen); 
+WKE_API void WKE_CALL wkePostURLW(wkeWebView wkeView,const wchar_t* url, const char* postData, int postLen);
 
-typedef void (*wkeAlertBoxCallback)(wkeWebView webView, void* param, const wkeString msg);
-WKE_API void wkeOnAlertBox(wkeWebView webView, wkeAlertBoxCallback callback, void* callbackParam);
+WKE_API void WKE_CALL wkeLoadHTML(wkeWebView webView, const utf8* html);
+WKE_API void WKE_CALL wkeLoadHTMLW(wkeWebView webView, const wchar_t* html);
 
-typedef bool (*wkeConfirmBoxCallback)(wkeWebView webView, void* param, const wkeString msg);
-WKE_API void wkeOnConfirmBox(wkeWebView webView, wkeConfirmBoxCallback callback, void* callbackParam);
+WKE_API void WKE_CALL wkeLoadFile(wkeWebView webView, const utf8* filename);
+WKE_API void WKE_CALL wkeLoadFileW(wkeWebView webView, const wchar_t* filename);
 
-typedef bool (*wkePromptBoxCallback)(wkeWebView webView, void* param, const wkeString msg, const wkeString defaultResult, wkeString result);
-WKE_API void wkeOnPromptBox(wkeWebView webView, wkePromptBoxCallback callback, void* callbackParam);
+WKE_API void WKE_CALL wkeLoad(wkeWebView webView, const utf8* str);
+WKE_API void WKE_CALL wkeLoadW(wkeWebView webView, const wchar_t* str);
+
+WKE_API bool WKE_CALL wkeIsLoading(wkeWebView webView);
+WKE_API bool WKE_CALL wkeIsLoadingSucceeded(wkeWebView webView);
+WKE_API bool WKE_CALL wkeIsLoadingFailed(wkeWebView webView);
+WKE_API bool WKE_CALL wkeIsLoadingCompleted(wkeWebView webView);
+WKE_API bool WKE_CALL wkeIsDocumentReady(wkeWebView webView);
+WKE_API void WKE_CALL wkeStopLoading(wkeWebView webView);
+WKE_API void WKE_CALL wkeReload(wkeWebView webView);
+
+WKE_API const utf8* WKE_CALL wkeGetTitle(wkeWebView webView);
+WKE_API const wchar_t* WKE_CALL wkeGetTitleW(wkeWebView webView);
+
+WKE_API void WKE_CALL wkeResize(wkeWebView webView, int w, int h);
+WKE_API int WKE_CALL wkeGetWidth(wkeWebView webView);
+WKE_API int WKE_CALL wkeGetHeight(wkeWebView webView);
+WKE_API int WKE_CALL wkeGetContentWidth(wkeWebView webView);
+WKE_API int WKE_CALL wkeGetContentHeight(wkeWebView webView);
+
+WKE_API void WKE_CALL wkeSetDirty(wkeWebView webView, bool dirty);
+WKE_API bool WKE_CALL wkeIsDirty(wkeWebView webView);
+WKE_API void WKE_CALL wkeAddDirtyArea(wkeWebView webView, int x, int y, int w, int h);
+WKE_API void WKE_CALL wkeLayoutIfNeeded(wkeWebView webView);
+WKE_API void WKE_CALL wkePaint(wkeWebView webView, void* bits,int bufWid, int bufHei, int xDst, int yDst, int w, int h, int xSrc, int ySrc, bool bCopyAlpha);
+WKE_API void WKE_CALL wkePaint2(wkeWebView webView, void* bits,int pitch);
+WKE_API void WKE_CALL wkeRepaintIfNeeded(wkeWebView webView);
+WKE_API HDC WKE_CALL wkeGetViewDC(wkeWebView webView);
+
+WKE_API bool WKE_CALL wkeCanGoBack(wkeWebView webView);
+WKE_API bool WKE_CALL wkeGoBack(wkeWebView webView);
+WKE_API bool WKE_CALL wkeCanGoForward(wkeWebView webView);
+WKE_API bool WKE_CALL wkeGoForward(wkeWebView webView);
+
+WKE_API void WKE_CALL wkeEditorSelectAll(wkeWebView webView);
+WKE_API void WKE_CALL wkeEditorCopy(wkeWebView webView);
+WKE_API void WKE_CALL wkeEditorCut(wkeWebView webView);
+WKE_API void WKE_CALL wkeEditorPaste(wkeWebView webView);
+WKE_API void WKE_CALL wkeEditorDelete(wkeWebView webView);
+
+WKE_API const wchar_t* WKE_CALL wkeGetCookieW(wkeWebView webView);
+WKE_API const utf8* WKE_CALL wkeGetCookie(wkeWebView webView);
+WKE_API void WKE_CALL wkeSetCookieEnabled(wkeWebView webView, bool enable);
+WKE_API bool WKE_CALL wkeIsCookieEnabled(wkeWebView webView);
+
+WKE_API void WKE_CALL wkeSetMediaVolume(wkeWebView webView, float volume);
+WKE_API float WKE_CALL wkeGetMediaVolume(wkeWebView webView);
+
+WKE_API bool WKE_CALL wkeFireMouseEvent(wkeWebView webView, unsigned int message, int x, int y, unsigned int flags);
+WKE_API bool WKE_CALL wkeFireContextMenuEvent(wkeWebView webView, int x, int y, unsigned int flags);
+WKE_API bool WKE_CALL wkeFireMouseWheelEvent(wkeWebView webView, int x, int y, int delta, unsigned int flags);
+WKE_API bool WKE_CALL wkeFireKeyUpEvent(wkeWebView webView, unsigned int virtualKeyCode, unsigned int flags, bool systemKey);
+WKE_API bool WKE_CALL wkeFireKeyDownEvent(wkeWebView webView, unsigned int virtualKeyCode, unsigned int flags, bool systemKey);
+WKE_API bool WKE_CALL wkeFireKeyPressEvent(wkeWebView webView, unsigned int charCode, unsigned int flags, bool systemKey);
+
+WKE_API void WKE_CALL wkeSetFocus(wkeWebView webView);
+WKE_API void WKE_CALL wkeKillFocus(wkeWebView webView);
+
+WKE_API wkeRect WKE_CALL wkeGetCaretRect(wkeWebView webView);
+
+WKE_API jsValue WKE_CALL wkeRunJS(wkeWebView webView, const utf8* script);
+WKE_API jsValue WKE_CALL wkeRunJSW(wkeWebView webView, const wchar_t* script);
+
+WKE_API jsExecState WKE_CALL wkeGlobalExec(wkeWebView webView);
+
+WKE_API void WKE_CALL wkeSleep(wkeWebView webView);
+WKE_API void WKE_CALL wkeWake(wkeWebView webView);
+WKE_API bool WKE_CALL wkeIsAwake(wkeWebView webView);
+
+WKE_API void WKE_CALL wkeSetZoomFactor(wkeWebView webView, float factor);
+WKE_API float WKE_CALL wkeGetZoomFactor(wkeWebView webView);
+
+WKE_API void WKE_CALL wkeSetEditable(wkeWebView webView, bool editable);
+
+WKE_API void WKE_CALL wkeSetHostWindow(wkeWebView webWindow, HWND hostWindow);
+WKE_API HWND WKE_CALL wkeGetHostWindow(wkeWebView webWindow);
+
+
+WKE_API const utf8* WKE_CALL wkeGetString(const wkeString string);
+WKE_API const wchar_t* WKE_CALL wkeGetStringW(const wkeString string);
+
+WKE_API void WKE_CALL wkeSetString(wkeString string, const utf8* str, size_t len);
+WKE_API void WKE_CALL wkeSetStringW(wkeString string, const wchar_t* str, size_t len);
+
+typedef void (WKE_CALL *wkeTitleChangedCallback)(wkeWebView webView, void* param, const wkeString title);
+WKE_API void WKE_CALL wkeOnTitleChanged(wkeWebView webView, wkeTitleChangedCallback callback, void* callbackParam);
+
+typedef void (WKE_CALL *wkeURLChangedCallback)(wkeWebView webView, void* param, const wkeString url);
+WKE_API void WKE_CALL wkeOnURLChanged(wkeWebView webView, wkeURLChangedCallback callback, void* callbackParam);
+
+typedef void (WKE_CALL *wkePaintUpdatedCallback)(wkeWebView webView, void* param, const HDC hdc, int x, int y, int cx, int cy);
+WKE_API void WKE_CALL wkeOnPaintUpdated(wkeWebView webView, wkePaintUpdatedCallback callback, void* callbackParam);
+
+typedef void (WKE_CALL *wkeAlertBoxCallback)(wkeWebView webView, void* param, const wkeString msg);
+WKE_API void WKE_CALL wkeOnAlertBox(wkeWebView webView, wkeAlertBoxCallback callback, void* callbackParam);
+
+typedef bool (WKE_CALL *wkeConfirmBoxCallback)(wkeWebView webView, void* param, const wkeString msg);
+WKE_API void WKE_CALL wkeOnConfirmBox(wkeWebView webView, wkeConfirmBoxCallback callback, void* callbackParam);
+
+typedef bool (WKE_CALL *wkePromptBoxCallback)(wkeWebView webView, void* param, const wkeString msg, const wkeString defaultResult, wkeString result);
+WKE_API void WKE_CALL wkeOnPromptBox(wkeWebView webView, wkePromptBoxCallback callback, void* callbackParam);
 
 
 typedef enum 
@@ -352,8 +351,8 @@ typedef struct
 
 } wkeConsoleMessage;
 
-typedef void (*wkeConsoleMessageCallback)(wkeWebView webView, void* param, const wkeConsoleMessage* message);
-WKE_API void wkeOnConsoleMessage(wkeWebView webView, wkeConsoleMessageCallback callback, void* callbackParam);
+typedef void (WKE_CALL *wkeConsoleMessageCallback)(wkeWebView webView, void* param, const wkeConsoleMessage* message);
+WKE_API void WKE_CALL wkeOnConsoleMessage(wkeWebView webView, wkeConsoleMessageCallback callback, void* callbackParam);
 
 
 
@@ -368,8 +367,8 @@ typedef enum
 
 } wkeNavigationType;
 
-typedef bool (*wkeNavigationCallback)(wkeWebView webView, void* param, wkeNavigationType navigationType, const wkeString url);
-WKE_API void wkeOnNavigation(wkeWebView webView, wkeNavigationCallback callback, void* param);
+typedef bool (WKE_CALL *wkeNavigationCallback)(wkeWebView webView, void* param, wkeNavigationType navigationType, const wkeString url);
+WKE_API void WKE_CALL wkeOnNavigation(wkeWebView webView, wkeNavigationCallback callback, void* param);
 
 
 
@@ -390,12 +389,12 @@ typedef struct
 
 } wkeWindowFeatures;
 
-typedef wkeWebView (*wkeCreateViewCallback)(wkeWebView webView, void* param, wkeNavigationType navigationType, const wkeString url, const wkeWindowFeatures* windowFeatures);
-WKE_API void wkeOnCreateView(wkeWebView webView, wkeCreateViewCallback callback, void* param);
+typedef wkeWebView (WKE_CALL *wkeCreateViewCallback)(wkeWebView webView, void* param, wkeNavigationType navigationType, const wkeString url, const wkeWindowFeatures* windowFeatures);
+WKE_API void WKE_CALL wkeOnCreateView(wkeWebView webView, wkeCreateViewCallback callback, void* param);
 
 
-typedef void (*wkeDocumentReadyCallback)(wkeWebView webView, void* param);
-WKE_API void wkeOnDocumentReady(wkeWebView webView, wkeDocumentReadyCallback callback, void* param);
+typedef void (WKE_CALL *wkeDocumentReadyCallback)(wkeWebView webView, void* param);
+WKE_API void WKE_CALL wkeOnDocumentReady(wkeWebView webView, wkeDocumentReadyCallback callback, void* param);
 
 
 typedef enum
@@ -406,8 +405,8 @@ typedef enum
 
 } wkeLoadingResult;
 
-typedef void (*wkeLoadingFinishCallback)(wkeWebView webView, void* param, const wkeString url, wkeLoadingResult result, const wkeString failedReason);
-WKE_API void wkeOnLoadingFinish(wkeWebView webView, wkeLoadingFinishCallback callback, void* param);
+typedef void (WKE_CALL *wkeLoadingFinishCallback)(wkeWebView webView, void* param, const wkeString url, wkeLoadingResult result, const wkeString failedReason);
+WKE_API void WKE_CALL wkeOnLoadingFinish(wkeWebView webView, wkeLoadingFinishCallback callback, void* param);
 
 
 typedef enum
@@ -418,26 +417,26 @@ typedef enum
 
 } wkeWindowType;
 
-WKE_API wkeWebView wkeCreateWebWindow(wkeWindowType type, HWND parent, int x, int y, int width, int height);
-WKE_API void wkeDestroyWebWindow(wkeWebView webWindow);
-WKE_API HWND wkeGetWindowHandle(wkeWebView webWindow);
+WKE_API wkeWebView WKE_CALL wkeCreateWebWindow(wkeWindowType type, HWND parent, int x, int y, int width, int height);
+WKE_API void WKE_CALL wkeDestroyWebWindow(wkeWebView webWindow);
+WKE_API HWND WKE_CALL wkeGetWindowHandle(wkeWebView webWindow);
 
-typedef bool (*wkeWindowClosingCallback)(wkeWebView webWindow, void* param);
-WKE_API void wkeOnWindowClosing(wkeWebView webWindow, wkeWindowClosingCallback callback, void* param);
+typedef bool (WKE_CALL *wkeWindowClosingCallback)(wkeWebView webWindow, void* param);
+WKE_API void WKE_CALL wkeOnWindowClosing(wkeWebView webWindow, wkeWindowClosingCallback callback, void* param);
 
-typedef void (*wkeWindowDestroyCallback)(wkeWebView webWindow, void* param);
-WKE_API void wkeOnWindowDestroy(wkeWebView webWindow, wkeWindowDestroyCallback callback, void* param);
+typedef void (WKE_CALL *wkeWindowDestroyCallback)(wkeWebView webWindow, void* param);
+WKE_API void WKE_CALL wkeOnWindowDestroy(wkeWebView webWindow, wkeWindowDestroyCallback callback, void* param);
 
 
-WKE_API void wkeShowWindow(wkeWebView webWindow, bool show);
-WKE_API void wkeEnableWindow(wkeWebView webWindow, bool enable);
+WKE_API void WKE_CALL wkeShowWindow(wkeWebView webWindow, bool show);
+WKE_API void WKE_CALL wkeEnableWindow(wkeWebView webWindow, bool enable);
 
-WKE_API void wkeMoveWindow(wkeWebView webWindow, int x, int y, int width, int height);
-WKE_API void wkeMoveToCenter(wkeWebView webWindow);
-WKE_API void wkeResizeWindow(wkeWebView webWindow, int width, int height);
+WKE_API void WKE_CALL wkeMoveWindow(wkeWebView webWindow, int x, int y, int width, int height);
+WKE_API void WKE_CALL wkeMoveToCenter(wkeWebView webWindow);
+WKE_API void WKE_CALL wkeResizeWindow(wkeWebView webWindow, int width, int height);
 
-WKE_API void wkeSetWindowTitle(wkeWebView webWindow, const utf8* title);
-WKE_API void wkeSetWindowTitleW(wkeWebView webWindow, const wchar_t* title);
+WKE_API void WKE_CALL wkeSetWindowTitle(wkeWebView webWindow, const utf8* title);
+WKE_API void WKE_CALL wkeSetWindowTitleW(wkeWebView webWindow, const wchar_t* title);
 
 
 /***JavaScript Bind***/
@@ -455,55 +454,55 @@ typedef enum
 } jsType;
 
 
-WKE_API void jsBindFunction(const char* name, jsNativeFunction fn, unsigned int argCount);
-WKE_API void jsBindGetter(const char* name, jsNativeFunction fn); /*get property*/
-WKE_API void jsBindSetter(const char* name, jsNativeFunction fn); /*set property*/
+WKE_API void WKE_CALL jsBindFunction(const char* name, jsNativeFunction fn, unsigned int argCount);
+WKE_API void WKE_CALL jsBindGetter(const char* name, jsNativeFunction fn); /*get property*/
+WKE_API void WKE_CALL jsBindSetter(const char* name, jsNativeFunction fn); /*set property*/
 
-WKE_API int jsArgCount(jsExecState es);
-WKE_API jsType jsArgType(jsExecState es, int argIdx);
-WKE_API jsValue jsArg(jsExecState es, int argIdx);
+WKE_API int WKE_CALL jsArgCount(jsExecState es);
+WKE_API jsType WKE_CALL jsArgType(jsExecState es, int argIdx);
+WKE_API jsValue WKE_CALL jsArg(jsExecState es, int argIdx);
 
-WKE_API jsType jsTypeOf(jsValue v);
-WKE_API bool jsIsNumber(jsValue v);
-WKE_API bool jsIsString(jsValue v);
-WKE_API bool jsIsBoolean(jsValue v);
-WKE_API bool jsIsObject(jsValue v);
-WKE_API bool jsIsFunction(jsValue v);
-WKE_API bool jsIsUndefined(jsValue v);
-WKE_API bool jsIsNull(jsValue v);
-WKE_API bool jsIsArray(jsValue v);
-WKE_API bool jsIsTrue(jsValue v);
-WKE_API bool jsIsFalse(jsValue v);
+WKE_API jsType WKE_CALL jsTypeOf(jsValue v);
+WKE_API bool WKE_CALL jsIsNumber(jsValue v);
+WKE_API bool WKE_CALL jsIsString(jsValue v);
+WKE_API bool WKE_CALL jsIsBoolean(jsValue v);
+WKE_API bool WKE_CALL jsIsObject(jsValue v);
+WKE_API bool WKE_CALL jsIsFunction(jsValue v);
+WKE_API bool WKE_CALL jsIsUndefined(jsValue v);
+WKE_API bool WKE_CALL jsIsNull(jsValue v);
+WKE_API bool WKE_CALL jsIsArray(jsValue v);
+WKE_API bool WKE_CALL jsIsTrue(jsValue v);
+WKE_API bool WKE_CALL jsIsFalse(jsValue v);
 
-WKE_API int jsToInt(jsExecState es, jsValue v);
-WKE_API float jsToFloat(jsExecState es, jsValue v);
-WKE_API double jsToDouble(jsExecState es, jsValue v);
-WKE_API bool jsToBoolean(jsExecState es, jsValue v);
-WKE_API const utf8* jsToTempString(jsExecState es, jsValue v);
-WKE_API const wchar_t* jsToTempStringW(jsExecState es, jsValue v);
+WKE_API int WKE_CALL jsToInt(jsExecState es, jsValue v);
+WKE_API float WKE_CALL jsToFloat(jsExecState es, jsValue v);
+WKE_API double WKE_CALL jsToDouble(jsExecState es, jsValue v);
+WKE_API bool WKE_CALL jsToBoolean(jsExecState es, jsValue v);
+WKE_API const utf8* WKE_CALL jsToTempString(jsExecState es, jsValue v);
+WKE_API const wchar_t* WKE_CALL jsToTempStringW(jsExecState es, jsValue v);
 
-WKE_API jsValue jsInt(int n);
-WKE_API jsValue jsFloat(float f);
-WKE_API jsValue jsDouble(double d);
-WKE_API jsValue jsBoolean(bool b);
+WKE_API jsValue WKE_CALL jsInt(int n);
+WKE_API jsValue WKE_CALL jsFloat(float f);
+WKE_API jsValue WKE_CALL jsDouble(double d);
+WKE_API jsValue WKE_CALL jsBoolean(bool b);
 
-WKE_API jsValue jsUndefined();
-WKE_API jsValue jsNull();
-WKE_API jsValue jsTrue();
-WKE_API jsValue jsFalse();
+WKE_API jsValue WKE_CALL jsUndefined();
+WKE_API jsValue WKE_CALL jsNull();
+WKE_API jsValue WKE_CALL jsTrue();
+WKE_API jsValue WKE_CALL jsFalse();
 
-WKE_API jsValue jsString(jsExecState es, const utf8* str);
-WKE_API jsValue jsStringW(jsExecState es, const wchar_t* str);
-WKE_API jsValue jsEmptyObject(jsExecState es);
-WKE_API jsValue jsEmptyArray(jsExecState es);
+WKE_API jsValue WKE_CALL jsString(jsExecState es, const utf8* str);
+WKE_API jsValue WKE_CALL jsStringW(jsExecState es, const wchar_t* str);
+WKE_API jsValue WKE_CALL jsEmptyObject(jsExecState es);
+WKE_API jsValue WKE_CALL jsEmptyArray(jsExecState es);
 
 
 
 //cexer JS对象、函数绑定支持
-typedef jsValue (*jsGetPropertyCallback)(jsExecState es, jsValue object, const char* propertyName);
-typedef bool (*jsSetPropertyCallback)(jsExecState es, jsValue object, const char* propertyName, jsValue value);
-typedef jsValue (*jsCallAsFunctionCallback)(jsExecState es, jsValue object, jsValue* args, int argCount);
-typedef void (*jsFinalizeCallback)(struct tagjsData* data);
+typedef jsValue (WKE_CALL *jsGetPropertyCallback)(jsExecState es, jsValue object, const char* propertyName);
+typedef bool (WKE_CALL *jsSetPropertyCallback)(jsExecState es, jsValue object, const char* propertyName, jsValue value);
+typedef jsValue (WKE_CALL *jsCallAsFunctionCallback)(jsExecState es, jsValue object, jsValue* args, int argCount);
+typedef void (WKE_CALL *jsFinalizeCallback)(struct tagjsData* data);
 
 typedef struct tagjsData
 {
@@ -515,37 +514,37 @@ typedef struct tagjsData
 
 } jsData;
 
-WKE_API jsValue jsObject(jsExecState es, jsData* obj);
-WKE_API jsValue jsFunction(jsExecState es, jsData* obj);
-WKE_API jsData* jsGetData(jsExecState es, jsValue object);
+WKE_API jsValue WKE_CALL jsObject(jsExecState es, jsData* obj);
+WKE_API jsValue WKE_CALL jsFunction(jsExecState es, jsData* obj);
+WKE_API jsData* WKE_CALL jsGetData(jsExecState es, jsValue object);
 
-WKE_API jsValue jsGet(jsExecState es, jsValue object, const char* prop);
-WKE_API void   jsSet(jsExecState es, jsValue object, const char* prop, jsValue v);
+WKE_API jsValue WKE_CALL jsGet(jsExecState es, jsValue object, const char* prop);
+WKE_API void   WKE_CALL jsSet(jsExecState es, jsValue object, const char* prop, jsValue v);
 
-WKE_API jsValue jsGetAt(jsExecState es, jsValue object, int index);
-WKE_API void   jsSetAt(jsExecState es, jsValue object, int index, jsValue v);
+WKE_API jsValue WKE_CALL jsGetAt(jsExecState es, jsValue object, int index);
+WKE_API void   WKE_CALL jsSetAt(jsExecState es, jsValue object, int index, jsValue v);
 
-WKE_API int     jsGetLength(jsExecState es, jsValue object);
-WKE_API void    jsSetLength(jsExecState es, jsValue object, int length);
+WKE_API int     WKE_CALL jsGetLength(jsExecState es, jsValue object);
+WKE_API void    WKE_CALL jsSetLength(jsExecState es, jsValue object, int length);
 
 
 
 //window object
-WKE_API jsValue jsGlobalObject(jsExecState es);
-WKE_API wkeWebView jsGetWebView(jsExecState es);
+WKE_API jsValue WKE_CALL jsGlobalObject(jsExecState es);
+WKE_API wkeWebView WKE_CALL jsGetWebView(jsExecState es);
 
-WKE_API jsValue jsEval(jsExecState es, const utf8* str);
-WKE_API jsValue jsEvalW(jsExecState es, const wchar_t* str);
+WKE_API jsValue WKE_CALL jsEval(jsExecState es, const utf8* str);
+WKE_API jsValue WKE_CALL jsEvalW(jsExecState es, const wchar_t* str);
 
-WKE_API jsValue jsCall(jsExecState es, jsValue func, jsValue thisObject, jsValue* args, int argCount);
-WKE_API jsValue jsCallGlobal(jsExecState es, jsValue func, jsValue* args, int argCount);
+WKE_API jsValue WKE_CALL jsCall(jsExecState es, jsValue func, jsValue thisObject, jsValue* args, int argCount);
+WKE_API jsValue WKE_CALL jsCallGlobal(jsExecState es, jsValue func, jsValue* args, int argCount);
 
-WKE_API jsValue jsGetGlobal(jsExecState es, const char* prop);
-WKE_API void    jsSetGlobal(jsExecState es, const char* prop, jsValue v);
+WKE_API jsValue WKE_CALL jsGetGlobal(jsExecState es, const char* prop);
+WKE_API void    WKE_CALL jsSetGlobal(jsExecState es, const char* prop, jsValue v);
 
 
 //garbage collect
-WKE_API void jsGC(); 
+WKE_API void   WKE_CALL  jsGC();